Change the BITVEC type to be of int and perform BitVec generation per bit
The unification routines try to match a BitVec of negative size to one of
positive size. The negative size is treated as a variable, so all BitVecs of
size -1 will be matched to the same length BitVec. There is still int to string
conversions because the length is parsed as a string and the unification
routines use an environment, env, of type (string * t).